Move overlay and moveable to moveable package
