+++ packages/moveable/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@googleforcreators/moveable",
+  "description": "The core Web Stories editor web application, which is platform-agnostic.",
+  "private": false,
+  "version": "0.0.1",
+  "author": {
+    "name": "Google",
+    "url": "https://creators.google/"
+  },
+  "license": "Apache-2.0",
+  "keywords": [
+    "web stories"
+  ],
+  "homepage": "https://github.com/googleforcreators/web-stories-wp/",
+  "repository": "git+https://github.com/GoogleForCreators/web-stories-wp.git",
+  "bugs": {
+    "url": "https://github.com/googleforcreators/web-stories-wp/issues"
+  },
+  "engines": {
+    "node": ">= 16",
+    "npm": ">= 7.3"
+  },
+  "type": "module",
+  "exports": {
+    "default": "./src/index.js"
+  },
+  "main": "./src/index.js",
+  "dependencies": {
+    "@googleforcreators/design-system": "*",
+    "@googleforcreators/react": "*",
+    "react-moveable": "^0.30.3",
+    "prop-types": "^15.7.2",
+    "styled-components": "^5.3.3"
+  },
+  "devDependencies": {
+    "prop-types": "^15.7.2"
+  }
+}
