+++ packages/moveable/src/overlay/index.js
@@ -23,7 +23,7 @@ import { forwardRef, useContext, createPortal } from '@googleforcreators/react';
 /**
  * Internal dependencies
  */
-import pointerEventsCss from '../../utils/pointerEventsCss';
+import { pointerEventsCss } from '../utils';
 import Context from './context';
 
 const Wrapper = styled.div`
